Create teleportspeedjump.lua
Dear users You can insert this into any injector and in this way. You can insert it there using the injector. You can insert this script into any injector. And what does it allow there is only it allows E E to run quickly, that is, to change with Speed ​​of your movement to change the strength of the jump, and also you can there by entering a special place the name of the player to teleport to him, but keep in mind, write it correctly and to teleport to him, and then start on enter and automatically. Then you will lose his name there, and you will teleport to him